# [build] Don't build "real" debug DSOs