We always build both release and debug binaries but the purpose of the
Debug build is not to be debuggable/unoptimized/etc in the native sense
but rather to contain code specific to and required by the managed Debug
builds.

So far, the Debug native builds have been built without any
optimizations thus making our Debug apps slower than they really need to
be. This commit changes the compiler optimization to `-O2` for the Debug
build, putting it on par with the Release builds, as far as optimization
is concerned.

This should make our Debug apps a bit faster.

The changes aren't dramatic, but there's a slight improvement:

= Before

Date (UTC): **3/25/2020 8:23:52 AM**
App configuration: **Debug**

Xamarin.Android
  - Version: **10.3.99-129**
  - Branch: **master**
  - Commit: **5099550c47d1ec39179a8cb7b05ca37fa3e61b74**

Device
  - Model: **Pixel 3 XL**
  - Native architecture: **arm64-v8a**
  - SDK version: **29**

= After

Date (UTC): **3/25/2020 9:48:59 AM**
App configuration: **Debug**

Xamarin.Android
  - Version: **10.3.99-130**
  - Branch: **always-optimize-native-code**
  - Commit: **00ac073bb855e2e88ed1b383745b3aa89b7a39cd**

Device
  - Model: **Pixel 3 XL**
  - Native architecture: **arm64-v8a**
  - SDK version: **29**

= Comparison

== Displayed time

| Before  | After   | Change    |
| ------- | ------- | --------- |
| 995.500 | 988.200 | -0.73% 🠅  |

== Native to managed transition time

| Before | After  | Change    |
| ------ | ------ | --------- |
| 50.187 | 50.164 | -0.04% 🠅  |

== Total native init time

| Before | After  | Change    |
| ------ | ------ | --------- |
| 89.513 | 87.692 | -2.03% 🠅  |