@brendanzagaeski 2020-03-23
*Draft release note*

It sounds like this change might give a nice improvement in startup times for Debug configuration apps. When you get a chance, you can add a draft release note for it to `Documentation/release-notes/4443.md` based on the [app startup performance template](https://github.com/xamarin/xamarin-android/blob/master/Documentation/release-notes/README.md#template-for-build-time-or-app-performance-improvement). Feel free to pick and choose parts of the template to suit this PR.

Also feel free to hold off on adding the note for as long as needed, if for example the data on the estimated time improvement won't be available until a little later in the PR.

Thanks in advance!