@grendello 2020-03-25
I found out that Android NDK was overriding our `-O2` and updated our code to disable the override. However, I don't think the results deserve mention in release notes. We are slightly faster in the native part, as you can see, but it doesn't translate much to the overall Displayed time speedup, so for the end users this information would just be inconsequential noise :)