@jonpryor 2020-03-26
Current draft release notes

---

`libmono-android.debug.so` and `libmono-android.release.so` differ
not in how easy they are to "debug" *themselves*, but rather in which
*app* configuration they should be included in:

  * `libmono-android.debug.so` is for "Debug" configuration apps, and
    supports loading assemblies from *outside* the `.apk`, e.g. for
    commercial Fast Deployment support.  It ***MUST NOT*** be used in
    "real" apps deployed to an App Store.

  * `libmono-android.release.so` is for "Release" configuration apps,
    and *only* loads assemblies from within the `.apk`.  (Even if the
    `.__override__` directory used by `libmono-android.debug.so`
    exists, *it will not be used*.)  This is the library that
    ***MUST*** be used in "real" apps deployed to an App Store.

Which is to say, the meaning of "debug" used in these file names does
not match the "conventional" use of "debug".

Unfortunately, we didn't inform `cmake` of this.  As such,
`libmono-android.debug.so` was also built with debug symbols and all
compiler optimizations *disabled* (`-O0`), which is unfortunate as
we want "Debug" configuration apps to start as quickly as possible,
to improve the inner dev loop.

Inform `cmake` and the NDK build system that we really want `-O2` to
be used to compile `libmono-android.debug.so`, which is the same set
of optimizations used for `libmono-android.release.so`.

This should make our Debug apps a bit faster.

The changes aren't dramatic, but there's a slight improvement:

Device:

  - Model: Pixel 3 XL
  - Native architecture: arm64-v8a
  - SDK version: 29

| Event                             | Before (ms) | After (ms) |     Δ     |
|-----------------------------------|------------:|-----------:|:---------:|
|                    Displayed time |     995.500 |    988.200 | -0.73% ⬆ |
| Native to managed transition time |      50.187 |     50.164 | -0.04% ⬆ |
|            Total native init time |      89.513 |     87.692 | -2.03% ⬆ |